refactor(server): Merge all error related code into one file
